Conflicts:
	LitebaseSDK/src/java/litebase/ResultSet.java
	LitebaseSDK/src/java/litebase/SQLSelectStatement.java
	LitebaseSDK/src/java/litebase/Table.java

+++ LitebaseSDK/src/java/litebase/SQLSelectClause.java
@@ -96,7 +96,7 @@ class SQLSelectClause
          String tableName;
          String[] columnNames;
          int[] columnHashes;
-         int[] columnTypes;
+         short[] columnTypes;
          int[] columnSizes;
          int pos = 0,
              count = 0,
